+++ extensions/1password/src/v8/utils.ts
@@ -4,21 +4,54 @@ import { execFileSync } from "child_process";
 import { existsSync } from "fs";
 import { useEffect, useState } from "react";
 
-import { CategoryName } from "./types";
+import { CategoryName, Item } from "./types";
+
+export type ActionID = "open-in-1password" | "open-in-browser" | "copy-username" | "copy-password";
 
 export type Preferences = {
   cliPath: string;
   version: "v7" | "v8";
+  primaryAction: ActionID;
+  secondaryAction: ActionID;
 };
 
 export const cache = new Cache();
 
+const preferences = getPreferenceValues<Preferences>();
+
 export const CLI_PATH =
-  getPreferenceValues<Preferences>().cliPath ||
-  ["/usr/local/bin/op", "/opt/homebrew/bin/op"].find((path) => existsSync(path));
+  preferences.cliPath || ["/usr/local/bin/op", "/opt/homebrew/bin/op"].find((path) => existsSync(path));
 export const CATEGORIES_CACHE_NAME = "@categories";
 export const ITEMS_CACHE_NAME = "@items";
 export const ACCOUNT_CACHE_NAME = "@account";
+
+export function hrefToOpenInBrowser(item: Item): string | undefined {
+  if (item.category === "LOGIN") {
+    return item.urls?.find((url) => url.primary)?.href;
+  } else {
+    return undefined;
+  }
+}
+
+export function actionsForItem(item: Item): ActionID[] {
+  if (item.category === "LOGIN") {
+    // user-configured primary action first, then secondary action,
+    // then all the actions in the default order,
+    // with duplicates removed
+    return [
+      ...new Set<ActionID>([
+        preferences.primaryAction,
+        preferences.secondaryAction,
+        "open-in-1password",
+        "open-in-browser",
+        "copy-username",
+        "copy-password",
+      ]),
+    ];
+  } else {
+    return ["open-in-1password"];
+  }
+}
 
 export function op(args: string[]) {
   if (CLI_PATH) {