+++ extensions/bitwarden/src/components/searchVault/actions/index.ts
@@ -1,0 +1,9 @@
+export { default as CopyPasswordAction } from "./CopyPasswordAction";
+export { default as PastePasswordAction } from "./PastePasswordAction";
+export { default as CopyTotpAction } from "./CopyTotpAction";
+export { default as ShowSecureNoteAction } from "./ShowSecureNoteAction";
+export { default as SearchCommonActions } from "./CommonActions";
+export { default as CopyUsernameAction } from "./CopyUsernameAction";
+export { default as CopyTextFieldsActions } from "./CopyTextFieldsActions";
+export { default as OpenUrlInBrowserAction } from "./OpenUrlInBrowserAction";
+export { default as ShowCardDetailsAction } from "./ShowCardDetailsAction";