+++ extensions/easydict/src/translation/microsoft/bing.ts
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-09-17 10:35
  * @lastEditor: tisfeng
- * @lastEditTime: 2023-01-17 16:22
+ * @lastEditTime: 2023-03-17 10:12
  * @fileName: bing.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.