+++ extensions/lastpass/src/index.tsx
@@ -1,15 +1,28 @@
-import { Action, ActionPanel, getPreferenceValues, Icon, List } from "@raycast/api";
+import { Action, ActionPanel, getPreferenceValues, Icon, List, LocalStorage } from "@raycast/api";
 import { useEffect, useState } from "react";
 import { lastPass } from "./cli";
 import { EmptyListView, ErrorDetails, ListItem } from "./components";
 
+type SyncRate = "0" | "86400000" | "604800000";
 interface Preferences {
   email: string;
   password: string;
+  syncRate: SyncRate;
 }
 
+const calculateSyncState = async (syncRate: SyncRate): Promise<"now" | "no"> => {
+  const localStorageKey = "lastpass-sync-timestamp";
+  const currentTimestamp = Date.now();
+  const lastSyncTimestamp = (await LocalStorage.getItem<number>(localStorageKey)) || currentTimestamp;
+  await LocalStorage.setItem(localStorageKey, currentTimestamp);
+  const timestampDiff = parseInt(syncRate, 10);
+  const isSyncNow = currentTimestamp - lastSyncTimestamp > timestampDiff;
+  return isSyncNow ? "now" : "no";
+};
+
 export default function Command() {
-  const { email, password } = getPreferenceValues<Preferences>();
+  const { email, password, syncRate } = getPreferenceValues<Preferences>();
+
   const api = lastPass(email, password);
   const [isLoading, setIsLoading] = useState(true);
   const [accounts, setAccounts] = useState<
@@ -25,7 +38,7 @@ export default function Command() {
           await api.login();
         }
 
-        const accounts = await api.list();
+        const accounts = await calculateSyncState(syncRate).then((sync) => api.list({ sync }));
         setAccounts(accounts);
         setIsLoading(false);
       } catch (error) {
@@ -47,7 +60,7 @@ export default function Command() {
           icon={Icon.ArrowClockwise}
           title="Manual Sync"
           shortcut={{ modifiers: ["cmd"], key: "s" }}
-          onAction={() => api.list({ sync: "now" }).then(setAccounts, setError)}
+          onAction={() => api.export({ sync: "now" }).then(setAccounts, setError)}
         />
       </ActionPanel.Section>
     </ActionPanel>
@@ -58,7 +71,12 @@ export default function Command() {
       {!accounts.length ? (
         <EmptyListView />
       ) : (
-        accounts.map((account) => <ListItem {...account} getDetails={() => api.show(account.id)} />)
+        accounts.map((account) => (
+          <ListItem
+            {...account}
+            getDetails={() => calculateSyncState(syncRate).then((sync) => api.show(account.id, { sync }))}
+          />
+        ))
       )}
     </List>
   );