+++ extensions/timers/src/start2MinuteTimer.ts
@@ -1,7 +1,8 @@
 import { closeMainWindow } from "@raycast/api";
-import { startTimer } from "./timerUtils";
+import { checkForOverlyLoudAlert, startTimer } from "./timerUtils";
 
 export default async () => {
+  if (!checkForOverlyLoudAlert()) return;
   await closeMainWindow();
   startTimer(60 * 2, "2 Minute Timer");
 };