+++ extensions/xcode/src/services/apple-developer-documentation.service.ts
@@ -43,7 +43,7 @@ export class AppleDeveloperDocumentationService {
     // For each Entry
     for (const entry of entries) {
       // Update URL
-      entry.url = [AppleDeveloperDocumentationService.hostUrl, entry.url].join("/");
+      entry.url = [AppleDeveloperDocumentationService.hostUrl, entry.url].join(entry.url.startsWith("/") ? "" : "/");
     }
     // Return Documentation Entries
     return entries;